+++ extras/src/scalatui/extras/package.scala
@@ -1,0 +1,3 @@
+package scalatui
+
+package object extras